Rename client to agent
