+++ remote/agent/agent.go
@@ -1,4 +1,4 @@
-package client
+package agent
 
 import (
 	"context"
